# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Volume-Manager service stores volumes and storage and allows to manage them",
    "title": "Volume-Manager API"
  },
  "paths": {
    "/admin/volumes": {
      "get": {
        "tags": [
          "Volumes"
        ],
        "summary": "Get all volumes (admin only).",
        "operationId": "GetAllVolumes",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "$ref": "#/parameters/Filters"
          },
          {
            "$ref": "#/parameters/PageNum"
          },
          {
            "$ref": "#/parameters/PerPageLimit"
          }
        ],
        "responses": {
          "200": {
            "description": "volumes response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Volume"
              }
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "delete": {
        "tags": [
          "Volumes"
        ],
        "summary": "Delete all user volumes (admin only).",
        "operationId": "DeleteAllUserVolumes",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          }
        ],
        "responses": {
          "200": {
            "description": "volumes deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/namespaces/{ns_id}/volumes": {
      "get": {
        "tags": [
          "Volumes"
        ],
        "summary": "Get user volumes.",
        "operationId": "GetUserVolumes",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "$ref": "#/parameters/Filters"
          },
          {
            "$ref": "#/parameters/NamespaceID"
          }
        ],
        "responses": {
          "200": {
            "description": "volumes response",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Volume"
              }
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "post": {
        "description": "Should be chosen first storage, where free space allows to create volume with provided capacity.",
        "tags": [
          "Volumes"
        ],
        "summary": "Create Volume for User by Tariff.",
        "operationId": "CreateVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "$ref": "#/parameters/NamespaceID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VolumeCreateRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "volume created"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/namespaces/{ns_id}/volumes/{label}": {
      "get": {
        "tags": [
          "Volumes"
        ],
        "summary": "Get volume.",
        "operationId": "GetVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "$ref": "#/parameters/NamespaceID"
          }
        ],
        "responses": {
          "200": {
            "description": "volume response",
            "schema": {
              "$ref": "#/definitions/Volume"
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "put": {
        "tags": [
          "Volumes"
        ],
        "summary": "Resize volume.",
        "operationId": "ResizeVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "$ref": "#/parameters/NamespaceID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VolumeResizeRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "volume resized"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "delete": {
        "tags": [
          "Volumes"
        ],
        "summary": "Delete volume.",
        "operationId": "DeleteVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "$ref": "#/parameters/NamespaceID"
          }
        ],
        "responses": {
          "200": {
            "description": "volume deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/namespaces/{ns_id}/volumes/{label}/rename": {
      "put": {
        "tags": [
          "Volumes"
        ],
        "summary": "Rename volume.",
        "operationId": "RenameVolume",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "$ref": "#/parameters/NamespaceID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/VolumeRenameRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "volume renamed"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/storages": {
      "get": {
        "tags": [
          "Storages"
        ],
        "summary": "Get storage list.",
        "operationId": "GetStorages",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          }
        ],
        "responses": {
          "200": {
            "description": "storages list",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Storage"
              }
            }
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "post": {
        "tags": [
          "Storages"
        ],
        "summary": "Create storage.",
        "operationId": "CreateStorage",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Storage"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "storage created"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    },
    "/storages/{name}": {
      "put": {
        "tags": [
          "Storages"
        ],
        "summary": "Update storage.",
        "operationId": "UpdateStorage",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/UpdateStorageRequest"
            }
          },
          {
            "type": "string",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "storage updated"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      },
      "delete": {
        "tags": [
          "Storages"
        ],
        "summary": "Delete storage.",
        "operationId": "DeleteStorage",
        "parameters": [
          {
            "$ref": "#/parameters/UserIDHeader"
          },
          {
            "$ref": "#/parameters/UserRoleHeader"
          },
          {
            "$ref": "#/parameters/SubstitutedUserID"
          },
          {
            "type": "string",
            "name": "name",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "202": {
            "description": "storage deleted"
          },
          "default": {
            "$ref": "#/responses/error"
          }
        }
      }
    }
  },
  "definitions": {
    "AccessLevel": {
      "type": "string",
      "x-go-package": "git.containerum.net/ch/volume-manager/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "CreateVolume": {
      "$ref": "vendor/github.com/containerum/kube-client/swagger.json#/definitions/CreateVolume"
    },
...
    "Err": {
      "description": "Err -- standard serializable API error\nMessage -- constant error message:\n+ \"invalid username\"\n+ \"quota exceeded\"\n+ \"validation error\"\n...etc...\nID -- unique error identification code\nDetails -- optional context error messages kinda\n+ \"field 'Replicas' must be non-zero value\"\n+ \"not enough tights to feed gopher\"\n+ \"resource 'God' does't exist\"\nFields -- optional set of key-value pairs",
      "type": "object",
      "properties": {
        "details": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Details"
        },
        "fields": {
          "$ref": "#/definitions/Fields"
        },
        "id": {
          "$ref": "#/definitions/ErrID"
        },
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "status_http": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "StatusHTTP"
        }
      },
      "x-go-package": "git.containerum.net/ch/volume-manager/vendor/github.com/containerum/cherry"
    },
    "ErrID": {
      "description": "ErrID -- represents unique error ID",
      "type": "object",
      "properties": {
        "kind": {
          "$ref": "#/definitions/ErrKind"
        },
        "sid": {
          "$ref": "#/definitions/ErrSID"
        }
      },
      "x-go-package": "git.containerum.net/ch/volume-manager/vendor/github.com/containerum/cherry"
    },
    "ErrKind": {
      "description": "ErrKind -- represents kind of error",
      "type": "integer",
      "format": "uint64",
      "x-go-package": "git.containerum.net/ch/volume-manager/vendor/github.com/containerum/cherry"
    },
    "ErrSID": {
      "description": "ErrSID -- represents service ID of error",
      "type": "string",
      "x-go-package": "git.containerum.net/ch/volume-manager/vendor/github.com/containerum/cherry"
    },
    "Fields": {
      "description": "Fields -- represents additional key-value fields of error",
      "type": "object",
      "additionalProperties": {
        "type": "string"
      },
      "x-go-package": "git.containerum.net/ch/volume-manager/vendor/github.com/containerum/cherry"
    },
...
    "PersistentVolumeAccessMode": {
      "type": "string",
      "x-go-package": "git.containerum.net/ch/volume-manager/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "ResourceUpdateName": {
      "$ref": "vendor/github.com/containerum/kube-client/swagger.json#/definitions/ResourceUpdateName"
    },
...
    "Storage": {
      "description": "Storage describes volumes storage",
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "ID"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Size"
        },
        "used": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Used"
        },
        "volumes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Volume"
          },
          "x-go-name": "Volumes"
        }
      },
      "x-go-package": "git.containerum.net/ch/volume-manager/pkg/models"
    },
...
    "UpdateStorageRequest": {
      "description": "UpdateStorageRequest represents request object for updating storage",
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "size": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "Used"
        }
      },
      "x-go-package": "git.containerum.net/ch/volume-manager/pkg/models"
    },
    "UserAccess": {
      "type": "object",
      "properties": {
        "access_level": {
          "$ref": "#/definitions/AccessLevel"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "git.containerum.net/ch/volume-manager/vendor/github.com/containerum/kube-client/pkg/model"
    },
...
    "Volume": {
      "description": "Volume -- volume representation",
      "type": "object",
      "properties": {
        "access": {
          "$ref": "#/definitions/AccessLevel"
        },
        "access_mode": {
          "$ref": "#/definitions/PersistentVolumeAccessMode"
        },
        "capacity": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "Capacity"
        },
        "created_at": {
          "type": "string",
          "x-go-name": "CreatedAt"
        },
        "name": {
          "type": "string",
          "x-go-name": "Name"
        },
        "owner": {
          "type": "string",
          "x-go-name": "Owner"
        },
        "owner_login": {
          "type": "string",
          "x-go-name": "OwnerLogin"
        },
        "storage_name": {
          "type": "string",
          "x-go-name": "StorageName"
        },
        "tariff_id": {
          "type": "string",
          "x-go-name": "TariffID"
        },
        "users": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/UserAccess"
          },
          "x-go-name": "Users"
        }
      },
      "x-go-package": "git.containerum.net/ch/volume-manager/vendor/github.com/containerum/kube-client/pkg/model"
    },
    "VolumeCreateRequest": {
      "description": "VolumeCreateRequest is a request object for creating volume",
      "$ref": "#/definitions/CreateVolume"
    },
    "VolumeRenameRequest": {
      "description": "VolumeRenameRequest is a request object for renaming volume",
      "$ref": "#/definitions/ResourceUpdateName"
    },
    "VolumeResizeRequest": {
      "description": "VolumeResizeRequest contains parameters for changing volume size",
      "type": "object",
      "properties": {
        "tariff_id": {
          "type": "string",
          "format": "uuid",
          "x-go-name": "TariffID"
        }
      },
      "x-go-package": "git.containerum.net/ch/volume-manager/pkg/models"
    },
...
  },
  "parameters": {
    "Filters": {
      "type": "string",
      "description": "A set of filters separated with comma. See \"dao\" package for more information.",
      "name": "filter",
      "in": "query"
    },
    "NamespaceID": {
      "type": "string",
      "format": "uuid",
      "description": "Namespace ID",
      "name": "ns_id",
      "in": "path",
      "required": true
    },
    "PageNum": {
      "minimum": 0,
      "type": "integer",
      "name": "page",
      "in": "query"
    },
    "PerPageLimit": {
      "minimum": 0,
      "type": "integer",
      "name": "per_page",
      "in": "query"
    },
    "SubstitutedUserID": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/SubstitutedUserID"
    },
    "UserIDHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserIDHeader"
    },
    "UserRoleHeader": {
      "$ref": "vendor/github.com/containerum/utils/httputil/swagger.json#/parameters/UserRoleHeader"
    }
  },
  "responses": {
    "error": {
      "description": "cherry error",
      "schema": {
        "$ref": "vendor/github.com/containerum/cherry/swagger.json#/definitions/Err"
      }
    }
  }
}